Created a GO server module with proper directories for each functionality
