+++ hub/web_app/server/go.mod
@@ -1,0 +1,5 @@
+module server
+
+go 1.23.6
+
+require github.com/golang-jwt/jwt/v4 v4.5.1
